requirements.txt
opencv-python
Pillow
numpy
streamlit
tensorflow
scikit-learn

